[LOG2LINES] Double-safety as workaround 64bit issue

+++ sdk/tools/log2lines/image.c
@@ -44,7 +44,7 @@ find_offset(void *data, size_t offset)
 
     for (i = 0; i < symbols; i++)
     {
-        if (Entries[i].Address > offset)
+        if (Entries[i].Address > (unsigned int)offset)
         {
             if (!i--)
                 return NULL;
